trying to fix LotteryEndToEndCancelLotto

+++ SadnaSrc/SadnaSrc/StoreCenter/StoreDL.cs
@@ -337,7 +337,7 @@ namespace SadnaSrc.StoreCenter
                 while (dbReader.Read())
                 {
                     LotteryTicket lottery = new LotteryTicket(dbReader.GetString(0), dbReader.GetString(1),
-                        dbReader.GetInt32(2), dbReader.GetInt32(3), dbReader.GetDouble(4), dbReader.GetInt32(6));
+                        dbReader.GetDouble(2), dbReader.GetDouble(3), dbReader.GetDouble(4), dbReader.GetInt32(6));
                     lottery.myStatus = EnumStringConverter.GetLotteryStatusString(dbReader.GetString(5));
                     result.AddLast(lottery);
                 }
@@ -738,7 +738,7 @@ namespace SadnaSrc.StoreCenter
                 {
                     lotteryManagement = new LotterySaleManagmentTicket(dbReader.GetString(0), dbReader.GetString(4), P,
                         DateTime.Parse(dbReader.GetString(5)), DateTime.Parse(dbReader.GetString(6)));
-                    lotteryManagement.TotalMoneyPayed = dbReader.GetInt32(3);
+                    lotteryManagement.TotalMoneyPayed = dbReader.GetDouble(3);
 
                     lotteryManagement.IsActive = (Boolean.Parse(dbReader.GetString(7)));
                 }
